requirements.txt
pillow>=10.0.0
imageio>=2.31.0
lottie[all]>=0.7.0
